requirements.txt
aiohttp
lxml_html_clean
Flask==2.2.2
gunicorn==20.1.0
werkzeug==2.2.2
aioshutil
aiofiles
anitopy
apscheduler
dnspython
motor==2.5.1
feedparser
psutil
pycountry
pymongo==3.12.3
python-dotenv
tgcrypto-pyrofork
torrentp==0.1.7
git+https://github.com/kaif-00z/html-telegraph-poster
uvloop
#https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
pyrofork
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

